+++ test/unit/openapi/batch-util-spec.js
@@ -90,33 +90,85 @@ describe('openapi batchUtil', () => {
 describe('batch building', () => {
 
     it('handles no requests', () => {
-        expect(batchBuild([], 'X', 'iitbank.com'))
-            .toEqual(multiline('--X--', ''));
+        expect(batchBuild([], 'iitbank.com'))
+            .toEqual({
+                body: multiline('--+--', ''),
+                boundary: '+',
+            });
     });
 
     it('handles one request', () => {
         expect(batchBuild(
                 [{ method: 'GET', url: 'openapi/sub' }],
-                'ABC', 'iitbank.com'))
-            .toEqual(multiline(
-                '--ABC',
-                'Content-Type:application/http; msgtype=request',
-                '',
-                'GET openapi/sub HTTP/1.1',
-                'X-Request-Id:0',
-                'Host:iitbank.com',
-                '',
-                '', // extra new line is important
-                '--ABC--',
-                ''));
+                'iitbank.com'))
+            .toEqual({
+                body: multiline(
+                        '--+',
+                        'Content-Type:application/http; msgtype=request',
+                        '',
+                        'GET openapi/sub HTTP/1.1',
+                        'X-Request-Id:0',
+                        'Host:iitbank.com',
+                        '',
+                        '', // extra new line is important
+                        '--+--',
+                        ''),
+                boundary: '+',
+            }
+            );
+    });
+
+    it('increases the boundary when it matches data', () => {
+        expect(batchBuild(
+            [{ method: 'POST', url: 'openapi/sub', data: '--+' }],
+            'iitbank.com'))
+            .toEqual({
+                body: multiline(
+                        '--++',
+                        'Content-Type:application/http; msgtype=request',
+                        '',
+                        'POST openapi/sub HTTP/1.1',
+                        'X-Request-Id:0',
+                        'Content-Type:application/json; charset=utf-8',
+                        'Host:iitbank.com',
+                        '',
+                        '--+',
+                        '--++--',
+                        ''),
+                boundary: '++',
+            }
+            );
+    });
+
+    it('increases the boundary when it matches data and the next character matches', () => {
+        expect(batchBuild(
+            [{ method: 'POST', url: 'openapi/sub', data: '--++' }],
+            'iitbank.com'))
+            .toEqual({
+                body: multiline(
+                        '--+-',
+                        'Content-Type:application/http; msgtype=request',
+                        '',
+                        'POST openapi/sub HTTP/1.1',
+                        'X-Request-Id:0',
+                        'Content-Type:application/json; charset=utf-8',
+                        'Host:iitbank.com',
+                        '',
+                        '--++',
+                        '--+---',
+                        ''),
+                boundary: '+-',
+            }
+            );
     });
 
     it('puts headers into the batch', () => {
         expect(batchBuild(
                 [{ method: 'GET', url: 'openapi/sub', headers: { 'X-Auth-Request': 'Me' } }],
-                'ABC', 'iitbank.com'))
-            .toEqual(multiline(
-                '--ABC',
+                'iitbank.com'))
+            .toEqual({
+                body: multiline(
+                '--+',
                 'Content-Type:application/http; msgtype=request',
                 '',
                 'GET openapi/sub HTTP/1.1',
@@ -125,16 +177,20 @@ describe('batch building', () => {
                 'Host:iitbank.com',
                 '',
                 '',
-                '--ABC--',
-                ''));
+                '--+--',
+                ''),
+                boundary: '+',
+            }
+            );
     });
 
     it('adds content-type for POST/PUT', () => {
         expect(batchBuild(
                 [{ method: 'POST', data: 'data', url: 'openapi/sub' }],
-                'ABC', 'iitbank.com'))
-            .toEqual(multiline(
-                '--ABC',
+                'iitbank.com'))
+            .toEqual({
+                body: multiline(
+                '--+',
                 'Content-Type:application/http; msgtype=request',
                 '',
                 'POST openapi/sub HTTP/1.1',
@@ -143,14 +199,18 @@ describe('batch building', () => {
                 'Host:iitbank.com',
                 '',
                 'data',
-                '--ABC--',
-                ''));
+                '--+--',
+                ''),
+                boundary: '+',
+            }
+            );
 
         expect(batchBuild(
                 [{ method: 'PUT', data: 'data', url: 'openapi/sub' }],
-                'ABC', 'iitbank.com'))
-            .toEqual(multiline(
-                '--ABC',
+                'iitbank.com'))
+            .toEqual({
+                body: multiline(
+                '--+',
                 'Content-Type:application/http; msgtype=request',
                 '',
                 'PUT openapi/sub HTTP/1.1',
@@ -159,8 +219,11 @@ describe('batch building', () => {
                 'Host:iitbank.com',
                 '',
                 'data',
-                '--ABC--',
-                ''));
+                '--+--',
+                ''),
+                boundary: '+',
+            }
+            );
     });
 
     it('handles multiple requests', () => {
@@ -168,9 +231,10 @@ describe('batch building', () => {
             [{ method: 'POST', data: 'postdata', url: 'openapi/sub' },
                 { method: 'PUT', data: 'putdata', url: 'openapi/bus' },
                 { method: 'GET', url: 'openapi/usb' }],
-                'ABC', 'iitbank.com'))
-            .toEqual(multiline(
-                '--ABC',
+                'iitbank.com'))
+            .toEqual({
+                body: multiline(
+                '--+',
                 'Content-Type:application/http; msgtype=request',
                 '',
                 'POST openapi/sub HTTP/1.1',
@@ -179,7 +243,7 @@ describe('batch building', () => {
                 'Host:iitbank.com',
                 '',
                 'postdata',
-                '--ABC',
+                '--+',
                 'Content-Type:application/http; msgtype=request',
                 '',
                 'PUT openapi/bus HTTP/1.1',
@@ -188,7 +252,7 @@ describe('batch building', () => {
                 'Host:iitbank.com',
                 '',
                 'putdata',
-                '--ABC',
+                '--+',
                 'Content-Type:application/http; msgtype=request',
                 '',
                 'GET openapi/usb HTTP/1.1',
@@ -196,7 +260,10 @@ describe('batch building', () => {
                 'Host:iitbank.com',
                 '',
                 '',
-                '--ABC--',
-                ''));
+                '--+--',
+                ''),
+                boundary: '+',
+            }
+            );
     });
 });